requirements.txt
asgiref==3.2.10
Django==3.1.14
django-debug-toolbar==2.2.1
django-environ==0.4.5
pytz==2019.3
sqlparse==0.4.4
djangorestframework==3.11.0

